requirements.txt
astroid==3.3.10
black==25.1.0
click==8.2.0
dill==0.4.0
flake8==7.2.0
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy==1.15.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.13.0
pyflakes==3.3.2
pygame==2.6.1
pylint==3.3.7
pytest==8.3.5
tomlkit==0.13.2
typing_extensions==4.13.2
pygame>=2.5.0
strawberry-graphql>=0.234.0
uvicorn>=0.24.0
fastapi>=0.104.0
